indexing_service/requirements.txt
kafka-python-ng
typesense
typing-extensions
requests
